Preserve selections when reopening EditSettingActivity
